@@ -11,9 +11,10 @@
     "@testing-library/user-event": "^13.5.0",
     "bootstrap": "^5.1.3",
     "gh-pages": "^3.2.3",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.2.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.5",
     "react-scripts": "5.0.0",
     "scrollmagic": "^2.0.8",
     "semantic-ui-react": "^2.1.2",
